+++ highs/ipm/hipo/factorhighs/Swaps.cpp
@@ -21,7 +21,7 @@ void permuteWithSwaps(double* x, const Int64* swaps, Int64 n, bool reverse) {
 }
 
 void swapCols(char uplo, Int64 n, double* A, Int64 lda, Int64 i, Int64 j, Int64* swaps,
-              Int64* sign, DataCollector& data) {
+              Int* sign, DataCollector& data) {
   // Exchange rows/cols i and j of symmetric matrix A
 
   // make sure that i < j
